Fixed ProcessorAccess check in the owner condition
