+++ entity-processor/src/main/java/com/fincity/saas/entity/processor/service/base/BaseProcessorService.java
@@ -4,6 +4,7 @@ import com.fincity.nocode.reactor.util.FlatMapUtil;
 import com.fincity.saas.commons.exeception.GenericException;
 import com.fincity.saas.entity.processor.dao.base.BaseProcessorDAO;
 import com.fincity.saas.entity.processor.dto.base.BaseProcessorDto;
+import com.fincity.saas.entity.processor.model.common.Identity;
 import com.fincity.saas.entity.processor.model.common.ProcessorAccess;
 import com.fincity.saas.entity.processor.service.ProcessorMessageResourceService;
 import java.util.Map;
@@ -51,6 +52,37 @@ public abstract class BaseProcessorService<
 
             return super.create(cEntity);
         });
+    }
+
+    public Mono<D> readIdentityWithOwnerAccess(Identity identity) {
+        return this.hasAccess().flatMap(access -> this.readIdentityWithOwnerAccess(access, identity));
+    }
+
+    public Mono<D> readIdentityWithOwnerAccess(ProcessorAccess access, Identity identity) {
+
+        if (identity == null || identity.isNull()) return this.identityMissingError();
+
+        return identity.isCode()
+                ? this.readByCode(access, identity.getCode())
+                        .flatMap(ticket -> this.checkUserAccess(access, ticket))
+                        .switchIfEmpty(this.msgService.throwMessage(
+                                msg -> new GenericException(HttpStatus.BAD_REQUEST, msg),
+                                ProcessorMessageResourceService.IDENTITY_WRONG,
+                                this.getEntityName(),
+                                identity.getCode()))
+                : this.readById(access, identity.getULongId())
+                        .flatMap(ticket -> this.checkUserAccess(access, ticket))
+                        .switchIfEmpty(this.msgService.throwMessage(
+                                msg -> new GenericException(HttpStatus.BAD_REQUEST, msg),
+                                ProcessorMessageResourceService.IDENTITY_WRONG,
+                                this.getEntityName(),
+                                identity.getId()));
+    }
+
+    public Mono<D> checkUserAccess(ProcessorAccess access, D entity) {
+        ULong accessUser = entity.getAccessUser();
+
+        return (accessUser != null && access.getSubOrg().contains(accessUser)) ? Mono.just(entity) : Mono.empty();
     }
 
     @Override
